Update Modify Order and Patient Reports to reuse existing components
