+++ frontend/src/components/addOrder/AddOrder.js
@@ -27,7 +27,7 @@ const AddOrder = (props) => {
 
   const componentMounted = useRef(false);
 
-  const { orderFormValues, setOrderFormValues, samples, error } = props;
+  const {orderFormValues, setOrderFormValues, samples, error ,isModifyOrder} = props;
   const [otherSamplingVisible, setOtherSamplingVisible] = useState(false);
   const [providers, setProviders] = useState([]);
   const [paymentOptions, setPaymentOptions] = useState([]);
@@ -261,6 +261,12 @@ const AddOrder = (props) => {
   };
 
   function handleLabNo(e, rawVal) {
+   if(isModifyOrder){
+    setOrderFormValues({
+      ...orderFormValues,
+      newAccessionNumber: e?.target?.value,
+    });
+   }else {
     setOrderFormValues({
       ...orderFormValues,
       sampleOrderItems: {
@@ -268,6 +274,7 @@ const AddOrder = (props) => {
         labNo: rawVal ? rawVal : e?.target?.value,
       },
     });
+   }
     setNotificationVisible(false);
   }
 
@@ -356,13 +363,21 @@ const AddOrder = (props) => {
 
   function fetchGeneratedAccessionNo(res) {
     if (res.status) {
-      setOrderFormValues({
-        ...orderFormValues,
-        sampleOrderItems: {
-          ...orderFormValues.sampleOrderItems,
-          labNo: res.body,
-        },
-      });
+      if(isModifyOrder){
+        setOrderFormValues({
+          ...orderFormValues,
+          newAccessionNumber: res.body,
+        });
+      }else{
+        setOrderFormValues({
+          ...orderFormValues,
+          sampleOrderItems: {
+            ...orderFormValues.sampleOrderItems,
+            labNo: res.body,
+          },
+        });
+      }
+
       setNotificationVisible(false);
     }
   }
@@ -408,6 +423,14 @@ const AddOrder = (props) => {
               value={orderFormValues.sampleOrderItems.externalOrderNumber}
             />
           )}
+          { isModifyOrder  && (
+              <h5>
+                {" "}
+                <FormattedMessage id="sample.label.labnumber" />:{" "}
+                {orderFormValues.accessionNumber}
+              </h5>
+              )
+            }
           <div className="formInlineDiv">
             <div className="inputText">
               <CustomLabNumberInput
@@ -415,7 +438,7 @@ const AddOrder = (props) => {
                 placeholder={intl.formatMessage({
                   id: "input.placeholder.labNo",
                 })}
-                value={orderFormValues.sampleOrderItems.labNo}
+                value={isModifyOrder? orderFormValues.newAccessionNumber : orderFormValues.sampleOrderItems.labNo}
                 onMouseLeave={handleLabNoValidation}
                 onChange={handleLabNo}
                 onKeyPress={handleKeyPress}
