+++ frontend/src/components/formModel/innitialValues/PatientStatusReportFormValues.js
@@ -1,12 +1,5 @@
 const PatientStatusReportFormValues = {
-  labNumber: "",
-  patientId: "",
   selectedPatientId: "",
-  lastName: "",
-  firstName: "",
-  dateOfBirth: "",
-  guid: "",
-  gender: "",
   referringSiteId: "",
   referringSiteName: "",
   startDate: "",
