+++ frontend/src/components/modifyOrder/EditSample.js
@@ -19,14 +19,14 @@ import {
 } from "@carbon/react";
 import { Add } from "@carbon/react/icons";
 import { getFromOpenElisServer } from "../utils/Utils";
-import EditSampleType from "./EditSampleType";
+import SampleType from "../addOrder/SampleType";
 import { FormattedMessage, useIntl } from "react-intl";
 import {
   OrderCurrentTestsHeaders,
   OrderPossibleTestsHeaders,
 } from "../data/orderCurrentTestsHeaders";
 const EditSample = (props) => {
-  const { samples, setSamples, orderFormValues, setOrderFormValues } = props;
+  const { samples, setSamples, orderFormValues, setOrderFormValues ,error} = props;
 
   const componentMounted = useRef(false);
 
@@ -486,12 +486,18 @@ const EditSample = (props) => {
                 <Link href="#" onClick={(e) => handleRemoveSample(e, sample)}>
                   {<FormattedMessage id="sample.remove.action" />}
                 </Link>
-                <EditSampleType
+                <SampleType
                   index={i}
                   rejectSampleReasons={rejectSampleReasons}
                   removeSample={removeSample}
                   sample={sample}
+                  setSample={(newSample) => {
+                    let newSamples = [...samples];
+                    newSamples[i] = newSample;
+                    setSamples(newSamples);
+                  }}
                   sampleTypeObject={sampleTypeObject}
+                  error={error}
                 />
               </div>
             );
