+++ frontend/src/components/reports/common/ReportByDateCSV.js
@@ -11,11 +11,11 @@ import {
   Loading,
 } from "@carbon/react";
 import { FormattedMessage, useIntl } from "react-intl";
-import "../../../Style.css";
-import { AlertDialog } from "../../../common/CustomNotification";
-import CustomDatePicker from "../../../common/CustomDatePicker";
-import config from "../../../../config.json";
-import { encodeDate, getFromOpenElisServer } from "../../../utils/Utils";
+import "../../Style.css";
+import { AlertDialog } from "../../common/CustomNotification";
+import CustomDatePicker from "../../common/CustomDatePicker";
+import config from "../../../config.json";
+import { encodeDate, getFromOpenElisServer } from "../../utils/Utils";
 
 
 const ReportByDateCSV = (props) => {
