+++ frontend/src/components/reports/common/ReportByID.js
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import { FormattedMessage, injectIntl, useIntl } from "react-intl";
 import { Form, Grid, Column, Section, Button, Loading } from "@carbon/react";
-import CustomLabNumberInput from "../../../common/CustomLabNumberInput";
-import { AlertDialog } from "../../../common/CustomNotification";
-import config from "../../../../config.json";
+import CustomLabNumberInput from "../../common/CustomLabNumberInput";
+import { AlertDialog } from "../../common/CustomNotification";
+import config from "../../../config.json";
 
 function ReportByID(props) {
   const intl = useIntl();
